+++ node/resolvers/items.ts
@@ -1,23 +1,50 @@
+import { Segment } from '@vtex/api'
 import { map } from 'bluebird'
+import crypto from 'crypto'
 
 import { SearchGraphQL } from '../clients/searchGraphQL'
 import { fixImageUrl } from '../utils/image'
 import { getNewOrderForm } from './orderForm'
 
-const getVariations = (skuId: string, skuList: any[]) => {
+const hashMD5 = (text: string) => {
+  const hash = crypto.createHash('md5')
+  return hash.update(text).digest('hex')
+}
+
+const toSpecificationIOMessage =
+  async (field: string, segment: Segment, content: string, id: string) => ({
+    content,
+    from: await segment.getSegmentByToken(null).then(x => x.cultureInfo),
+    id: `Specification-id.${id}::${field}`,
+  })
+
+const getVariations = (skuId: string, skuList: any[], segment: Segment) => {
   const matchedSku = skuList.find((sku: any) => sku.itemId === skuId)
   if (!matchedSku) {
     return []
   }
   return matchedSku.variations.map((variation: any) => ({
-    fieldName: variation.name,
-    fieldValues: variation.values,
+    fieldName: toSpecificationIOMessage(
+      'fieldName',
+      segment,
+      variation.name,
+      hashMD5(variation.name)
+    ),
+    fieldValues: variation.values.map(
+      (value: string) => toSpecificationIOMessage(
+        'fieldValue',
+        segment,
+        value,
+        hashMD5(value)
+      )
+    ),
   }))
 }
 
 export const adjustItems = (
   items: OrderFormItem[],
-  searchGraphQL: SearchGraphQL
+  searchGraphQL: SearchGraphQL,
+  segment: Segment
 ) =>
   map(items, async (item: OrderFormItem) => {
     const response = await searchGraphQL.product({
@@ -33,7 +60,7 @@ export const adjustItems = (
       ...item,
       imageUrl: fixImageUrl(item.imageUrl)!,
       name: product.productName,
-      skuSpecifications: getVariations(item.id, product.items),
+      skuSpecifications: getVariations(item.id, product.items, segment),
     }
   })
 
@@ -44,7 +71,7 @@ export const mutations = {
     ctx: Context
   ): Promise<OrderForm> => {
     const {
-      clients: { checkout, searchGraphQL },
+      clients: { checkout, searchGraphQL, segment },
       vtex: { orderFormId },
     } = ctx
 
@@ -72,6 +99,7 @@ export const mutations = {
       checkout,
       newOrderForm,
       searchGraphQL,
+      segment,
     })
   },
 }