+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `skuSpecifications` field now return translated name and values.
+
 ## [0.13.0] - 2019-09-24
 
 ### Added